// possbuzz-backend/src/product/dto/update-product.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreateProductDto } from './create-product.dto';

export class UpdateProductDto extends PartialType(CreateProductDto) {
  name?: string | undefined;
  sku?: string | undefined;
  price?: number | undefined;
  stockQty?: number | undefined;
}